+++ packages/fcl/src/app-utils/verify-signatures.js
@@ -61,13 +61,13 @@ const getVerifySignaturesScript = async (sig, opts) => {
       ? "verifyAccountProofSignatures"
       : "verifyUserSignatures"
 
-  let network = await getChainId()
+  let network = await getChainId(opts)
 
   let fclCryptoContract
 
   invariant(
     opts.fclCryptoContract || network === "testnet" || network === "mainnet",
-    "${verifyFunction}({ fclCryptoContract }) -- config.flow.network must be specified (testnet || mainnet) or contract address provided via opts.fclCryptoContract"
+    "${verifyFunction}({ fclCryptoContract }) -- fclCrypto contract address must be set for non-mainnet/testnet networks"
   )
 
   if (opts.fclCryptoContract) {