+++ packages/fcl/src/utils/chain-id/fetch-chain-id.js
@@ -1,0 +1,8 @@
+import * as sdk from "@onflow/sdk"
+
+export async function fetchChainId(opts = {}) {
+  const response = await sdk
+    .send([sdk.getNetworkParameters()], opts)
+    .then(sdk.decode)
+  return response.chainId
+}