+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @onflow/sdk
 
+## 1.2.1-alpha.0
+
+### Patch Changes
+
+- Updated dependencies [[`6193bd42`](https://github.com/onflow/fcl-js/commit/6193bd420371494a975f19fd001af6af012bd72f), [`fefb578a`](https://github.com/onflow/fcl-js/commit/fefb578ae96a112164c561fa3e0a490008d8a5f6), [`6fa3bdc4`](https://github.com/onflow/fcl-js/commit/6fa3bdc46dd077b57f501d802af5544393502285)]:
+  - @onflow/config@1.1.1-alpha.0
+  - @onflow/transport-http@1.7.0-alpha.0
+  - @onflow/util-logger@1.2.1-alpha.0
+
 ## 1.2.0
 
 ### Minor Changes